Move Promise to a separate file.
